requirements.txt
fastapi==0.110.0
pydantic==1.10.14
python-multipart==0.0.9
asyncpg==0.29.0
alembic==1.13.1
SQLAlchemy==2.0.29
redis==4.6.0
fastapi-cache2[redis]==0.2.1
PyJWT==2.8.0
bcrypt==4.1.2
pytz==2024.1
requests==2.31.0
